+++ src/Futhark/Analysis/Rephrase.hs
@@ -80,10 +80,10 @@ rephraseParam rephraser (Param name from) =
 
 -- | Rephrase a body.
 rephraseBody :: Monad m => Rephraser m from to -> Body from -> m (Body to)
-rephraseBody rephraser (Body rep bnds res) =
+rephraseBody rephraser (Body rep stms res) =
   Body
     <$> rephraseBodyDec rephraser rep
-    <*> (stmsFromList <$> mapM (rephraseStm rephraser) (stmsToList bnds))
+    <*> (stmsFromList <$> mapM (rephraseStm rephraser) (stmsToList stms))
     <*> pure res
 
 -- | Rephrase a lambda.