+++ src/Futhark/Optimise/Simplify/Rules/Loop.hs
@@ -122,9 +122,9 @@ hoistLoopInvariantMergeVariables vtable pat aux (merge, form, loopbody) = do
           isInvariant,
           -- Also do not remove the condition in a while-loop.
           not $ paramName mergeParam `nameIn` freeIn form =
-          let (bnd, explpat'') =
+          let (stm, explpat'') =
                 removeFromResult (mergeParam, mergeInit) explpat'
-           in ( maybe id (:) bnd $ (paramIdent mergeParam, mergeInit) : invariant,
+           in ( maybe id (:) stm $ (paramIdent mergeParam, mergeInit) : invariant,
                 explpat'',
                 merge',
                 resExps