+++ src/Futhark/Pass/ExtractKernels.hs
@@ -21,7 +21,7 @@
 -- @
 --   map
 --     map(f)
---     bnds_a...
+--     stms_a...
 --     map(g)
 -- @
 --
@@ -31,7 +31,7 @@
 --   map
 --     map(f)
 --   map
---     bnds_a
+--     stms_a
 --   map
 --     map(g)
 -- @
@@ -40,7 +40,7 @@
 --
 --  (0) it can be done without creating irregular arrays.
 --      Specifically, the size of the arrays created by @map(f)@, by
---      @map(g)@ and whatever is created by @bnds_a@ that is also used
+--      @map(g)@ and whatever is created by @stms_a@ that is also used
 --      in @map(g)@, must be invariant to the outermost loop.
 --
 --  (1) the maps are _balanced_.  That is, the functions @f@ and @g@
@@ -48,7 +48,7 @@
 --
 -- The advantage is that the map-nests containing @map(f)@ and
 -- @map(g)@ can now be trivially flattened at no cost, thus exposing
--- more parallelism.  Note that the @bnds_a@ map constitutes array
+-- more parallelism.  Note that the @stms_a@ map constitutes array
 -- expansion, which requires additional storage.
 --
 -- = Distributing Sequential Loops
@@ -249,20 +249,20 @@ type GPUStms = Stms Out.GPU
 
 transformBody :: KernelPath -> Body -> DistribM (Out.Body Out.GPU)
 transformBody path body = do
-  bnds <- transformStms path $ stmsToList $ bodyStms body
-  return $ mkBody bnds $ bodyResult body
+  stms <- transformStms path $ stmsToList $ bodyStms body
+  return $ mkBody stms $ bodyResult body
 
 transformStms :: KernelPath -> [Stm] -> DistribM GPUStms
 transformStms _ [] =
   return mempty
-transformStms path (bnd : bnds) =
-  sequentialisedUnbalancedStm bnd >>= \case
+transformStms path (stm : stms) =
+  sequentialisedUnbalancedStm stm >>= \case
     Nothing -> do
-      bnd' <- transformStm path bnd
-      inScopeOf bnd' $
-        (bnd' <>) <$> transformStms path bnds
-    Just bnds' ->
-      transformStms path $ stmsToList bnds' <> bnds
+      stm' <- transformStm path stm
+      inScopeOf stm' $
+        (stm' <>) <$> transformStms path stms
+    Just stms' ->
+      transformStms path $ stmsToList stms' <> stms
 
 unbalancedLambda :: Lambda -> Bool
 unbalancedLambda orig_lam =
@@ -401,8 +401,8 @@ transformStm path (Let res_pat aux (Op (Screma w arrs form)))
           | otherwise = comm,
     Just do_irwim <- irwim res_pat w comm' red_fun $ zip nes arrs = do
     types <- asksScope scopeForSOACs
-    (_, bnds) <- fst <$> runBuilderT (simplifyStms =<< collectStms_ (auxing aux do_irwim)) types
-    transformStms path $ stmsToList bnds
+    (_, stms) <- fst <$> runBuilderT (simplifyStms =<< collectStms_ (auxing aux do_irwim)) types
+    transformStms path $ stmsToList stms
 transformStm path (Let pat aux@(StmAux cs _ _) (Op (Screma w arrs form)))
   | Just (reds, map_lam) <- isRedomapSOAC form = do
     let paralleliseOuter = runBuilder_ $ do
@@ -578,8 +578,8 @@ transformStm _ (Let orig_pat (StmAux cs _ _) (Op (Hist w ops bucket_fun imgs))) 
     addStms =<< histKernel onLambda lvl orig_pat [] [] cs w ops bfun' imgs
   where
     onLambda = pure . soacsLambdaToGPU
-transformStm _ bnd =
-  runBuilder_ $ FOT.transformStmRecursively bnd
+transformStm _ stm =
+  runBuilder_ $ FOT.transformStmRecursively stm
 
 sufficientParallelism ::
   String ->
@@ -915,12 +915,12 @@ onInnerMap path maploop@(MapLoop pat aux w lam arrs) acc
           -- sequentialising.  This is only OK as long as further
           -- versioning does not take place down that branch (it currently
           -- does not).
-          (sequentialised_kernel, nestw_bnds) <- localScope extra_scope $ do
+          (sequentialised_kernel, nestw_stms) <- localScope extra_scope $ do
             let sequentialised_lam = soacsLambdaToGPU lam'
             constructKernel segThreadCapped nest' $ lambdaBody sequentialised_lam
 
           let outer_pat = loopNestingPat $ fst nest
-          (nestw_bnds <>)
+          (nestw_stms <>)
             <$> onMap'
               nest'
               path