+++ unittests/futhark_tests.hs
@@ -6,7 +6,7 @@ import qualified Futhark.IR.Mem.IxFunTests
 import qualified Futhark.IR.PrimitiveTests
 import qualified Futhark.IR.PropTests
 import qualified Futhark.IR.Syntax.CoreTests
-import qualified Futhark.Optimise.ReuseAllocations.GreedyColoringTests
+import qualified Futhark.Optimise.MemoryBlockMerging.GreedyColoringTests
 import qualified Futhark.Pkg.SolveTests
 import qualified Language.Futhark.SyntaxTests
 import Test.Tasty
@@ -22,7 +22,7 @@ allTests =
       Futhark.Pkg.SolveTests.tests,
       Futhark.IR.Mem.IxFunTests.tests,
       Futhark.IR.PrimitiveTests.tests,
-      Futhark.Optimise.ReuseAllocations.GreedyColoringTests.tests,
+      Futhark.Optimise.MemoryBlockMerging.GreedyColoringTests.tests,
       Futhark.Analysis.AlgSimplifyTests.tests
     ]